requirements.txt
flask
flask-sqlalchemy
flask-login
pymysql
cryptography
python-dotenv
werkzeug
opencv-python
numpy
deepface